+++ src/controllers/game_controller.cpp
@@ -28,12 +28,12 @@ void GameController::start() {
 
         cliViewController->writeOutput(String("In which harbor would you like to start?"));
 
-        for(int i = 0; i < world->getHarbors().getSize(); i++) {
+        for(size_t i = 0; i < world->getHarbors().getSize(); i++) {
             cliViewController->writeOutput(String() << i << String(": ") << world->getHarbors()[i]->getName());
         }
         try {
             const String input_string = cliViewController->getInput();
-            const int input = std::atoi(input_string.c_str());
+            const size_t input = std::atoi(input_string.c_str());
             if(input >= 0 && input < world->getHarbors().getSize()) {
                 world->getPlayer().getShip()->setDestination(world->getHarbors()[input]);
                 harborController->dockShip();
